divide InstructorCoursesPage into small components

+++ src/app/api/instructor/course/route.ts
@@ -10,7 +10,7 @@ export async function GET(req: NextRequest): Promise<NextResponse> {
   try {
     // Authenticate the user
     const session: Session | null = await getServerSession(authOptions);
-    if (!session?.user) {
+    if (!session?.user || session?.user?.currentRole !== 'instructor') {
       return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
     }
 
@@ -46,10 +46,27 @@ export async function GET(req: NextRequest): Promise<NextResponse> {
 // Update a course by course ID
 export async function PATCH(req: NextRequest): Promise<NextResponse> {
   try {
+    // Authenticate the user
+    const session: Session | null = await getServerSession(authOptions);
+    if (!session?.user || session?.user?.currentRole !== 'instructor') {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    }
+
     const searchParams: URLSearchParams = req.nextUrl.searchParams;
     const courseId: string | null = searchParams.get('courseId');
     await connectDB();
     const updatedContent: ICourse = await req.json();
+
+    // Check if the course exists
+    const course: ICourse | null = await Course.findById(courseId);
+    if (!course) {
+      return NextResponse.json({ error: 'Course not found' }, { status: 404 });
+    }
+
+    // Check if the course belongs to the instructor    
+    if (course.instructor.toString() !== session.user.id) {
+      return NextResponse.json({ error: 'The course does not belong to you, you cannot update it.' }, { status: 401 });
+    }
 
     const updatedResult: ICourse | null = await Course.findByIdAndUpdate(
       courseId,
@@ -70,9 +87,25 @@ export async function PATCH(req: NextRequest): Promise<NextResponse> {
 // Delete a course by course ID
 export async function DELETE(req: NextRequest): Promise<NextResponse> {
   try {
+    // Authenticate the user
+    const session: Session | null = await getServerSession(authOptions);
+    if (!session?.user || session?.user?.currentRole !== 'instructor') {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    }
+
     const searchParams: URLSearchParams = req.nextUrl.searchParams;
     const courseId: string | null = searchParams.get('courseId');
     await connectDB();
+
+    // Check if the course belongs to the instructor
+    const course: ICourse | null = await Course.findById(courseId);
+    if (!course) {
+      return NextResponse.json({ error: 'Course not found' }, { status: 404 });
+    }
+    if (course.instructor.toString() !== session.user.id) {
+      return NextResponse.json({ error: 'The course does not belong to you, you cannot delete it.' }, { status: 401 });
+    }
+
     const deletedResult: ICourse | null =
       await Course.findByIdAndDelete(courseId);
     return NextResponse.json({ deletedResult });
